Added Subscribe implementation
