--- aws_protocol_adaptor/client_device/cfg_aws.txt
@@ -1,26 +1,0 @@
-###############################################################################
-# Copyright 2010-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License").
-# You may not use this file except in compliance with the License.
-# A copy of the License is located at
-#
-#  http://aws.amazon.com/apache2.0
-#
-# or in the "license" file accompanying this file. This file is distributed
-# on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
-# express or implied. See the License for the specific language governing
-# permissions and limitations under the License.
-###############################################################################
-
-[message-broker]
-HostAddress = <YOUR IOT HOST ADDRESS>
-Port = 443
-ThingName = deepstrea-app-to-cloud
-ClientID = deepstrea-app-to-cloud
-RootCALocation = <DEEPSTREAM SDK PATH>/sources/libs/aws_protocol_adaptor/device_client/certs/root.ca.pem
-DeepstreamCertLocation = <DEEPSTREAM SDK PATH>/sources/libs/aws_protocol_adaptor/device_client/certs/certificatePem.cert.pem
-DeepstreamPrivateKeyLocation = <DEEPSTREAM SDK PATH>/sources/libs/aws_protocol_adaptor/device_client/certs/privateKey.private.key
-MQTTCommandTimeoutMS = 20000
-TLSHandshakeTimeoutMS = 20000
-SSLHostnameVerify = false
